requirements.txt
djangorestframework==3.14.0
django-cors-headers==3.13.0
Django==4.1.7
psycopg2-binary==2.9.5
python-dotenv==0.21.1
